+++ changelog.md
@@ -1,7 +1,7 @@
 ## v0.0.93
 * Refactor mediawiki-page-history computation + fix
 * Mediawiki-history: Handle dropping of user fields in labs views
-* Update mediawiki_history checker to historical values 
+* Update mediawiki_history checker to historical values
 * Update pageview definition to exclude non wiki sites
 * Add entropy UDAF to refinery-hive